add: report_ads

+++ app/common/index.js
@@ -21,10 +21,10 @@ export const APARTMENT_ROOM_TYPE = [
 
 export const PRICE_FORMAT = (price) => {
   if (parseInt(price) > 1000000 || parseInt(price) === 1000000) {
-    return (parseInt(price) / 1000000).toFixed(2) + I18n.t('PRICE_M')
+    return (parseFloat(price) / 1000000).toFixed(2) + I18n.t('PRICE_M')
   }
   if (parseInt(price) > 1000 || parseInt(price) === 1000) {
-    return (parseInt(price) / 1000).toFixed(2) + I18n.t('PRICE_K')
+    return (parseFloat(price) / 1000).toFixed(2) + I18n.t('PRICE_K')
   }  
 
   return parseFloat(price).toFixed(2)
